Add experience skills PUT functionality
Also, exposed bug in achievements functionality that made me decide that
it should be ripped out.

The issue I realized was that it was very easy to send invalid experiences
through either the POST or PUT endpoints, by including a skill on an
achievement but not on the experience itself. Basically, to add a skill
to an achievement correctly you would _always_ have to add it in two
places. The complexity around this added lots of tricky edge cases on
the backend code, which was a code smell in addition to being an easy
stumbling block for anyone using the API.

After some thought, I realized we're not actually going to use this
functionality, so I figured I'd just pull it out for now. However, if we
decide we need this functionality again in the future, we'll need to
build it differently to ensure that linkages between skills and
achievements are only specified in one place.